// utils/enums.ts
export enum EFieldTypes{
  checkbox = 1,
  selectbox = 2,
  textbox = 3,
  numericbox = 4
}

export enum EDriveInsertResult{
  driveInserted = 0,
  driveNotWipedInDb = 1,
}

export enum EJobReportType{
  r2 = 1,
  r2plus = 2,
  cod = 3
}

export enum EJobReportDetail{
  simple = 1,
  perDrive = 2,
}